rename text_spot to spot_text
