+++ mapreader/__init__.py
@@ -13,12 +13,12 @@ from mapreader.classify.classifier import ClassifierContainer
 from mapreader.classify import custom_models
 
 try:
-    from mapreader.text_spot.deepsolo_runner import DeepSoloRunner
+    from mapreader.spot_text.deepsolo_runner import DeepSoloRunner
 except ImportError:
     pass
 
 try:
-    from mapreader.text_spot.dptext_detr_runner import DPTextDETRRunner
+    from mapreader.spot_text.dptext_detr_runner import DPTextDETRRunner
 except ImportError:
     pass
