+++ mapreader/spot_text/deepsolo_runner.py
