+++ mapreader/spot_text/dptext_detr_runner.py
